requirements.txt
pandas>=2.0
numpy>=1.25
langdetect>=1.0.9
python-dateutil>=2.9